move Test1D.test_integrate1D to test_tools.py
